[Misc] Rename current breakpoint files to use snake_casing to match rest of repo
