+++ src/services/breakpoint/index.ts
@@ -7,5 +7,5 @@
  */
 
 export * from './breakpoint';
-export * from './currentEuiBreakpoint';
+export * from './current_breakpoint';
 export * from './useIsWithinBreakpoints';
